added new endpoint for getting user attention by date interval

+++ application/src/main/java/com/example/capstone/ws/dto/AverageAttentionDto.java
@@ -3,4 +3,4 @@ package com.example.capstone.ws.dto;
 import lombok.Builder;
 
 @Builder
-public record AverageAttentionDto(String identifier, Double percentage) { }
+public record AverageAttentionDto(String identifier, Long percentage) { }
